cvlogger/requirements.txt
openvino
opencv-python
numpy
pandas
scipy
mysql-connector-python
logging
python-dotenv